refactor(island): split the bubbles out, which the error banner pushed over the cap
`CommandCenterChatView`'s body reached 480 lines and SwiftLint's
`type_body_length` cap is 450. My own fault twice over: the error banner in the
previous commit was the eleven lines that crossed it, and I pushed that commit
having printed the lint count without reading it. CI would have caught this; it
should not have needed to.

`+Bubbles.swift` takes the rendering — the bubble, its markdown body and the
action row. That is the cohesive half: what is left is the panel, the prompt pill
and the wiring to the coordinator, and this is what one turn looks like.

Six members lost `private` so the extension can reach them, each marked
`// Extension-visible: +Bubbles` as CLAUDE.md requires — the marker is what tells
the next person why they are not private, so nobody "tidies" them back and breaks
the build.

Body is now 511 lines in the file and inside the cap. Both targets build, 1520
tests in 129 suites pass, SwiftFormat 0.62.1 and SwiftLint 0.65.0 clean at the
versions CI pins — checked by reading the output this time, not the count.
